+++ app/controllers/ProjectController.java
@@ -14,8 +14,10 @@ import play.data.Form;
 import play.db.jpa.Transactional;
 import play.mvc.Controller;
 import play.mvc.Result;
+import service.ActionsEnum;
 import service.Configuration;
 import service.ProjectConverter;
+import service.SecurityService;
 import views.data.MenuDto;
 import views.data.ProjectDto;
 import views.html.projects.projectDetail;
@@ -24,7 +26,15 @@ import views.html.projects.projects;
 import views.html.projects.projectsCreate;
 import views.html.projects.projectsEdit;
 import daos.impl.DAOs;
-
+import play.mvc.Security;
+/**
+ * Controller for actions related to Project
+ * @author Tomas Michalicka (<a href='mailto:tomas@michalicka.com'>tomas@michalicka.com</a>)
+ *
+ * @see Project
+ * @see ProjectDto
+ */
+@Security.Authenticated(Secured.class)
 public class ProjectController extends Controller{
 
 	
@@ -35,23 +45,35 @@ public class ProjectController extends Controller{
 	 */
 	@Transactional(readOnly=true)
 	public static Result showAll(Integer page) {
+		User user = SecurityService.fetchUser(session("authid"));
+		if (!SecurityService.hasAccess(user, ActionsEnum.PROJECT_SHOW_ALL)) {
+			return redirect(routes.Application.accessDenied());
+		}
 		if (page == null) {
 			page = 0;
+			Logger.debug("No page number is given. Setting 0 (first page).");
 		}
 		List<Project> proj= DAOs.getProjectDao().getAllProject(page * Configuration.PAGE_SIZE, Configuration.PAGE_SIZE);
 		
 		Integer totalProjects = DAOs.getProjectDao().getNumberOfProjects();
 		Integer numberOfPages = totalProjects % Configuration.PAGE_SIZE == 0 ? totalProjects/Configuration.PAGE_SIZE : totalProjects/Configuration.PAGE_SIZE + 1; 
-		return ok(projects.render(ProjectConverter.convertListToDto(proj), getMainMenu(), numberOfPages, page));
+		Logger.debug("Page with list of projects is shown. Number of projects id db is {}", totalProjects);
+		return ok(projects.render(ProjectConverter.convertListToDto(proj), getMainMenu(user), numberOfPages, page));
 	}
 	
 	/**
 	 * Action shows form for adding new project
 	 * @return
 	 */
+	@Transactional(readOnly=true)
 	public static Result create() {
-		Form<ProjectDto> userForm = Form.form(ProjectDto.class);
-		return ok(projectsCreate.render(userForm, getBackToListMenu()));
+		User user = SecurityService.fetchUser(session("authid"));
+		if (!SecurityService.hasAccess(user, ActionsEnum.PROJECT_CREATE)) {
+			return redirect(routes.Application.accessDenied());
+		}
+		Form<ProjectDto> projectForm = Form.form(ProjectDto.class);
+		Logger.debug("Page with form for creating new project is shown.");
+		return ok(projectsCreate.render(projectForm, getBackToListMenu(user)));
 	}
 	/**
 	 * Action saves new project
@@ -59,10 +81,15 @@ public class ProjectController extends Controller{
 	 */
 	@Transactional(readOnly=false)
 	public static Result saveNewProject() {
-		Form<ProjectDto> userForm = Form.form(ProjectDto.class).bindFromRequest();
-		Project newProject = ProjectConverter.convertToEntity(userForm.get());
+		User user = SecurityService.fetchUser(session("authid"));
+		if (!SecurityService.hasAccess(user, ActionsEnum.PROJECT_CREATE)) {
+			return redirect(routes.Application.accessDenied());
+		}
+		Form<ProjectDto> projectForm = Form.form(ProjectDto.class).bindFromRequest();
+		Project newProject = ProjectConverter.convertToEntity(projectForm.get());
 		newProject.setVisible(true);
 		DAOs.getProjectDao().create(newProject);
+		Logger.debug("Action for saving data of new project was called.");
 		return redirect(routes.ProjectController.showAll(0).absoluteURL(request()));
 	}
 	/**
@@ -72,11 +99,17 @@ public class ProjectController extends Controller{
 	 */
 	@Transactional(readOnly=false)
 	public static Result detail(Long projectId) {
+		User user = SecurityService.fetchUser(session("authid"));
+		if (!SecurityService.hasAccess(user, ActionsEnum.PROJECT_DETAIL)) {
+			return redirect(routes.Application.accessDenied());
+		}
 		Project project = DAOs.getProjectDao().findById(projectId);
 		if (project == null) {
+			Logger.info("Partner with id {} was not found, detail can not be shown.", projectId);
 			return redirect(routes.ProjectController.projectNotFound(projectId));
 		} else {
-			return ok(projectDetail.render(ProjectConverter.convertToDto(project), getBackToListMenu(), null, new ArrayList<User>()));
+			Logger.debug("Partner detail page is shown.");
+			return ok(projectDetail.render(ProjectConverter.convertToDto(project), getBackToListMenu(user), null, new ArrayList<User>()));
 		}
 	}
 	
@@ -87,8 +120,19 @@ public class ProjectController extends Controller{
 	 */
 	@Transactional(readOnly=false)
 	public static Result edit(Long projectId) {
-		Form<ProjectDto> projectForm = Form.form(ProjectDto.class).fill(ProjectConverter.convertToDto(DAOs.getProjectDao().findById(projectId)));
-		return ok(projectsEdit.render(projectForm, getBackToListMenu(), false));
+		User user = SecurityService.fetchUser(session("authid"));
+		if (!SecurityService.hasAccess(user, ActionsEnum.PROJECT_EDIT)) {
+			return redirect(routes.Application.accessDenied());
+		}		
+		//TODO tmichalicka - second level verification
+		ProjectDto dto = ProjectConverter.convertToDto(DAOs.getProjectDao().findById(projectId));
+		if (dto == null) {
+			Logger.info("Project with id {} was not found.", projectId);
+			return redirect(routes.ProjectController.projectNotFound(projectId));
+		}
+		Form<ProjectDto> projectForm = Form.form(ProjectDto.class).fill(dto);
+		Logger.debug("Page with form for editing project is shown. Edited project has id {} and name {}.", dto.getProjectId(), dto.getName());
+		return ok(projectsEdit.render(projectForm, getBackToListMenu(user), false));
 	}
 	
 	/**
@@ -98,10 +142,18 @@ public class ProjectController extends Controller{
 	 */
 	@Transactional(readOnly=false)
 	public static Result delete(Long projectId) {
+		User user = SecurityService.fetchUser(session("authid"));
+		if (!SecurityService.hasAccess(user, ActionsEnum.PROJECT_DELETE)) {
+			return redirect(routes.Application.accessDenied());
+		}	
 		Project project = DAOs.getProjectDao().findById(projectId);
 		if (project != null) {
 			DAOs.getProjectDao().delete(project);
+		} else {
+			Logger.info("Project with id was not found, delete opereation was not successful.", projectId);
+			return redirect(routes.ProjectController.projectNotFound(projectId));
 		}
+		Logger.debug("Delete operation was called to project with id {}.", projectId);
 		return redirect(routes.ProjectController.showAll(0).absoluteURL(request()));
 	}
 	
@@ -110,8 +162,10 @@ public class ProjectController extends Controller{
 	 * @param projectId
 	 * @return
 	 */
+	@Transactional(readOnly=true)
 	public static Result projectNotFound(Long projectId) {
-		return ok(projectNotFound.render(projectId, getBackToListMenu()));
+		Logger.debug("Project not found page is shown. Requested id was {}.", projectId);
+		return ok(projectNotFound.render(projectId, getBackToListMenu(SecurityService.fetchUser(session("authid")))));
 	}
 	
 	/**
@@ -121,11 +175,16 @@ public class ProjectController extends Controller{
 	 */
 	@Transactional(readOnly=false)
 	public static Result updateProject(Boolean forceNext) {
+		User user = SecurityService.fetchUser(session("authid"));
+		if (!SecurityService.hasAccess(user, ActionsEnum.PROJECT_EDIT)) {
+			return redirect(routes.Application.accessDenied());
+		}
 		Form<ProjectDto> projectForm = Form.form(ProjectDto.class).bindFromRequest();
 		Project project = ProjectConverter.convertToEntity(projectForm.get());
 		if (forceNext != null && forceNext) {
 			Project edited = DAOs.getProjectDao().findById(projectForm.get().getProjectId());
 			project.setVersion(edited.getVersion());
+			Logger.debug("Force rewrite project data is set. Previous modifications will be deleted.");
 		}
 		try {
 			project.setPartners(new HashSet<Partner>());
@@ -146,32 +205,40 @@ public class ProjectController extends Controller{
 			DAOs.getProjectDao().update(project);
 		} catch (OptimisticLockException e) {
 			Logger.info("Project {} was edited by another user. ", projectForm.get());
-			return ok(projectsEdit.render(projectForm, getBackToListMenu(), true));
+			return ok(projectsEdit.render(projectForm, getBackToListMenu(user), true));
 		}
+		Logger.debug("Project update operation was called.");
 		return redirect(routes.ProjectController.showAll(0));
 	}
-	
-	private static List<MenuDto> getBackToListMenu() {
+	/**
+	 * Method returns list of items to left side menu. This implementation returns one item - back to list
+	 * @return
+	 */
+	private static List<MenuDto> getBackToListMenu(User user) {
 		List<MenuDto> result = new ArrayList<MenuDto>();
-		
-		MenuDto back = new MenuDto();
-		back.setGlyphicon("triangle-left");
-		back.setLabel("Zpět na seznam projektů");
-		back.setUrl(routes.ProjectController.showAll(0).absoluteURL(request()));
-		result.add(back);
-		
+		if (SecurityService.hasAccess(user, ActionsEnum.PROJECT_SHOW_ALL)) {
+			MenuDto back = new MenuDto();
+			back.setGlyphicon("triangle-left");
+			back.setLabel("Zpět na seznam projektů");
+			back.setUrl(routes.ProjectController.showAll(0).absoluteURL(request()));
+			result.add(back);
+		}
 		return result;
 	}
 	
-	private static List<MenuDto> getMainMenu() {
+	/**
+	 * Method returns list of items to left side menu. This implementation returns one item - add new
+	 * @return
+	 */
+	private static List<MenuDto> getMainMenu(User user) {
 		List<MenuDto> result = new ArrayList<MenuDto>();
-		
-		MenuDto newProject = new MenuDto();
-		newProject.setGlyphicon("plus");
-		newProject.setLabel("Přidat projekt");
-		newProject.setUrl(routes.ProjectController.create().absoluteURL(request()));
-		result.add(newProject);
-		
+		if (SecurityService.hasAccess(user, ActionsEnum.PROJECT_CREATE)) {
+			MenuDto newProject = new MenuDto();
+			newProject.setGlyphicon("plus");
+			newProject.setLabel("Přidat projekt");
+			newProject.setUrl(routes.ProjectController.create().absoluteURL(request()));
+			result.add(newProject);
+		}
 		return result;		
 	}
